requirements.txt
certifi==2019.11.28
numpy==1.18.2
pandas==1.0.3
Pillow==7.0.0
protobuf==3.11.3
python-box==4.2.2
python-dateutil==2.8.1
pytz==2019.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
tensorboardX==2.0
toml==0.10.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.43.0
